+++ js/data.js
@@ -1,0 +1,45 @@
+const NAMES = [
+  'Viktor',
+  'Sonya',
+  'Nicola',
+  'Olga',
+  'Santiago'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const PHOTOS_COUNT = 25;
+
+const IdNumber = {
+  MIN: 1,
+  MAX: 25
+};
+
+const LikesCount = {
+  MIN: 15,
+  MAX: 200
+};
+
+const UrlNumber = {
+  MIN: 1,
+  MAX: 25
+};
+
+const CommentsCount = {
+  MIN: 0,
+  MAX: 30
+};
+
+const AvatarNumber = {
+  MIN: 1,
+  MAX: 6,
+};
+
+export {NAMES, MESSAGES, PHOTOS_COUNT, IdNumber, LikesCount, UrlNumber, CommentsCount, AvatarNumber}
